refactor(decomp): extract query-cache policy utilities → new crate (TD-DECOMP-29) STRUCTURAL ENABLER
Move 4 tractable query/cache files (per_category_policy 8t + mismatch_cost 8t +
category_classifier 14t + result_cache_gate 10t = 40 tests) into a new query-tier crate
proximadb-query-cache. Intra-module refs rewritten (4 sites, crate::query::cache::X→crate::X).
Root query/cache/mod.rs re-exports all 4 (pub use proximadb_query_cache::X).

STRUCTURAL ENABLER: creates the query-cache crate as a home for the remaining 5 coupled files
(query_result_cache 19t, plan_cache 14t, invalidation_coordinator 11t, invalidation 11t,
adaptive_cache 3t = 58 tests) to move into once their root-only deps extract:
- ExecutionResult (query::federated) blocks query_result_cache + invalidation
- PlanOutput (query::federated::optimizer) blocks plan_cache
- QueryCache (storage::cache) blocks invalidation_coordinator

40 tests out of root (920->960). Verified: standalone crate --all-targets EXIT 0 in ~9s
with RUSTC_WRAPPER=sccache. See TD-DECOMP-29.
