+++ crates/query/proximadb-query-cache/src/mismatch_cost.rs
