+++ crates/query/proximadb-query-cache/src/per_category_policy.rs
