+++ docs/10-quality/td/TD-DECOMP-29-extract-query-cache.adoc
@@ -1,0 +1,46 @@
+// Copyright (C) 2025 ProximaDB
+// SPDX-License-Identifier: Apache-2.0
+= TD-DECOMP-29: Extract query-cache policy utilities → new proximadb-query-cache crate (query tier) — STRUCTURAL ENABLER
+
+[cols="1,3", options="header"]
+|===
+| Field | Value
+| Status | Landing
+| Severity | Medium — structural enabler (partial module extraction; creates a home for the remaining coupled files)
+| Component | `crates/query/proximadb-query-cache`; root `src/query/cache/mod.rs`
+| Relates to | [[TD-DECOMP-1]] (recipe); [[TD-DECOMP-25]] (batch_group extracted from same module)
+|===
+
+== Why
+
+`src/query/cache/` has 98 tests across 9 files. batch_group (12 tests) was extracted earlier
+(TD-DECOMP-25). This TD extracts the **4 tractable files** (40 tests) whose only external refs are
+intra-module (to each other), creating the `proximadb-query-cache` crate as a **home** for the
+remaining coupled files to move into once their root-only deps resolve.
+
+== What moved (behavior-neutral)
+
+* `per_category_policy.rs` (8t, 0 crate::), `mismatch_cost.rs` (8t, 0 crate::),
+  `category_classifier.rs` (14t, 1 intra-module ref), `result_cache_gate.rs` (10t, 3 intra-module refs)
+  → `crates/query/proximadb-query-cache/src/`. Total: **40 tests**.
+* Intra-module refs rewritten: `crate::query::cache::X` → `crate::X` (4 sites, sibling-to-sibling).
+* Root `query/cache/mod.rs`: 4 `pub mod X;` → `pub use proximadb_query_cache::X;`.
+* New query-tier crate, dep `tokio` (workspace).
+
+== What stays root-side (and why)
+
+* `query_result_cache.rs` (19t) — depends on `ExecutionResult` (root-only, `query::federated`).
+* `plan_cache.rs` (14t) — depends on `PlanOutput` (root-only, `query::federated::optimizer`).
+* `invalidation_coordinator.rs` (11t, 14 crate:: refs) — depends on `QueryCache` (root-only, `storage::cache`).
+* `invalidation.rs` (11t) — test-only ref to `ExecutionResult`.
+* `adaptive_cache.rs` (3t) — depends on `QueryCacheKey` (in query_result_cache, root-staying).
+
+These move into `proximadb-query-cache` once their root-only deps extract to crates.
+
+== Follow-ups
+
+* **Running total once merged: ~960 tests out of root** (920 + 40).
+* Standalone crate `--all-targets` compiled in ~9s with `RUSTC_WRAPPER=sccache`.
+* Next: extract `ExecutionResult`/`PlanOutput` → query-fusion crate (federated types already
+  partially there); then `QueryCache` → a storage-cache crate. Each unblocks the corresponding
+  root-staying files.
